adding changes before the responsive feature
