+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\index.js":"1","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\App.js":"2","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\reportWebVitals.js":"3","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\Listing.jsx":"4"},{"size":517,"mtime":1608786353980,"results":"5","hashOfConfig":"6"},{"size":199,"mtime":1608728366808,"results":"7","hashOfConfig":"6"},{"size":375,"mtime":1608786353980,"results":"8","hashOfConfig":"6"},{"size":4024,"mtime":1608785308311,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"a0mhho",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\index.js",[],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\App.js",[],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\reportWebVitals.js",[],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\Listing.jsx",[]]
+[{"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\index.js":"1","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\App.js":"2","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\reportWebVitals.js":"3","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\Listing.jsx":"4","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\book_detail.js":"5"},{"size":517,"mtime":1608786353980,"results":"6","hashOfConfig":"7"},{"size":199,"mtime":1608728366808,"results":"8","hashOfConfig":"7"},{"size":375,"mtime":1608786353980,"results":"9","hashOfConfig":"7"},{"size":2354,"mtime":1608902339422,"results":"10","hashOfConfig":"7"},{"size":1998,"mtime":1608902377647,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"a0mhho",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"21","usedDeprecatedRules":"14"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\index.js",[],["24","25"],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\App.js",[],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\reportWebVitals.js",[],"G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\Listing.jsx",["26"],"import React, { Component, useState } from \"react\";\r\nimport \"./listing.css\";\r\n\r\nimport book_details from \"./book_detail\";\r\n\r\nclass Listing extends Component {\r\n\tconstructor(props) {\r\n\t\tsuper(props);\r\n\r\n\t\tthis.state = {\r\n\t\t\tinput: \"\",\r\n\t\t};\r\n\r\n\t\t// const [input, setInput] = useState(\"\");\r\n\t\t// const [table, setTable] = useState(null);\r\n\t}\r\n\r\n\tfilterTable = () => {\r\n\t\tif (this.state.input.length > 0) {\r\n\t\t\tlet tableElements = (\r\n\t\t\t\t<div className=\"book-not-found\">\r\n\t\t\t\t\t<h3>No such book available</h3>\r\n\t\t\t\t\t<h5>\r\n\t\t\t\t\t\tTip: Try searching for keywords matching the book name\r\n\t\t\t\t\t\tor author name\r\n\t\t\t\t\t</h5>\r\n\t\t\t\t</div>\r\n\t\t\t);\r\n\t\t\tconst tableHeaders = Object.keys(book_details[0]);\r\n\r\n\t\t\tlet matching_books = book_details.filter((book) => {\r\n\t\t\t\treturn (\r\n\t\t\t\t\tbook.author\r\n\t\t\t\t\t\t.toLowerCase()\r\n\t\t\t\t\t\t.includes(this.state.input.toLowerCase()) ||\r\n\t\t\t\t\tbook[\"book name\"]\r\n\t\t\t\t\t\t.toLowerCase()\r\n\t\t\t\t\t\t.includes(this.state.input.toLowerCase())\r\n\t\t\t\t);\r\n\t\t\t});\r\n\r\n\t\t\tif (matching_books.length > 0) {\r\n\t\t\t\ttableElements = (\r\n\t\t\t\t\t<table>\r\n\t\t\t\t\t\t<thead>\r\n\t\t\t\t\t\t\t<tr>\r\n\t\t\t\t\t\t\t\t{tableHeaders.map((header, index) => (\r\n\t\t\t\t\t\t\t\t\t<th key={index}>{header}</th>\r\n\t\t\t\t\t\t\t\t))}\r\n\t\t\t\t\t\t\t</tr>\r\n\t\t\t\t\t\t</thead>\r\n\r\n\t\t\t\t\t\t<tbody>\r\n\t\t\t\t\t\t\t{matching_books.map((book, index) => (\r\n\t\t\t\t\t\t\t\t<tr key={index}>\r\n\t\t\t\t\t\t\t\t\t<td>{book[\"id\"]}</td>\r\n\t\t\t\t\t\t\t\t\t<td>{book[\"vibhag id\"]}</td>\r\n\t\t\t\t\t\t\t\t\t<td>{book[\"book name\"]}</td>\r\n\t\t\t\t\t\t\t\t\t<td>{book[\"author\"]}</td>\r\n\t\t\t\t\t\t\t\t</tr>\r\n\t\t\t\t\t\t\t))}\r\n\t\t\t\t\t\t</tbody>\r\n\t\t\t\t\t</table>\r\n\t\t\t\t);\r\n\t\t\t}\r\n\r\n\t\t\t// return tableElements;\r\n\t\t\tdocument.querySelector(\".table\").innerHTML = tableElements;\r\n\t\t}\r\n\t};\r\n\r\n\trender() {\r\n\t\treturn (\r\n\t\t\t<div className=\"container\">\r\n\t\t\t\t<form className=\"form\">\r\n\t\t\t\t\t<input\r\n\t\t\t\t\t\tclassName=\"search-bar\"\r\n\t\t\t\t\t\ttype=\"text\"\r\n\t\t\t\t\t\tplaceholder=\"Search by Book name / Author name\"\r\n\t\t\t\t\t\tonInput={(event) => {\r\n\t\t\t\t\t\t\t// setInput(event.target.value);\r\n\t\t\t\t\t\t\tthis.setState({ input: event.target.value });\r\n\t\t\t\t\t\t}}\r\n\t\t\t\t\t/>\r\n\t\t\t\t\t<button\r\n\t\t\t\t\t\tclassName=\"search-button\"\r\n\t\t\t\t\t\ttype=\"submit\"\r\n\t\t\t\t\t\tonClick={(event) => {\r\n\t\t\t\t\t\t\tevent.preventDefault();\r\n\t\t\t\t\t\t\t// setTable(filterTable());\r\n\t\t\t\t\t\t\tthis.filterTable();\r\n\t\t\t\t\t\t}}\r\n\t\t\t\t\t>\r\n\t\t\t\t\t\tSearch\r\n\t\t\t\t\t</button>\r\n\t\t\t\t</form>\r\n\t\t\t\t{/* \r\n\t\t\t\t\t\tbuttonClicked ? table : \"\"\r\n\t\t\t\t\t*/}\r\n\t\t\t\t<div className=\"table\"></div>\r\n\t\t\t</div>\r\n\t\t);\r\n\t}\r\n}\r\n\r\nexport default Listing;\r\n","G:\\projects\\SaVaKhed Library\\New folder\\savakhed\\src\\components\\listing-page\\book_detail.js",[],{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","severity":1,"message":"32","line":1,"column":28,"nodeType":"33","messageId":"34","endLine":1,"endColumn":36},"no-native-reassign",["35"],"no-negated-in-lhs",["36"],"no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
