+++ src/components/listing-page/book_detail.js
@@ -1,0 +1,112 @@
+const book_details = [
+	{
+		id: "1",
+		"vibhag id": "11",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "2",
+		"vibhag id": "22",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "3",
+		"vibhag id": "33",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+	{
+		id: "4",
+		"vibhag id": "44",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "5",
+		"vibhag id": "55",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "6",
+		"vibhag id": "66",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+	{
+		id: "7",
+		"vibhag id": "77",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "8",
+		"vibhag id": "88",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "9",
+		"vibhag id": "99",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+	{
+		id: "1",
+		"vibhag id": "11",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "2",
+		"vibhag id": "22",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "3",
+		"vibhag id": "33",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+	{
+		id: "4",
+		"vibhag id": "44",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "5",
+		"vibhag id": "55",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "6",
+		"vibhag id": "66",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+	{
+		id: "7",
+		"vibhag id": "77",
+		"book name": "The Alchemist",
+		author: "Paulo Coelho",
+	},
+	{
+		id: "8",
+		"vibhag id": "88",
+		"book name": "Yoga",
+		author: "Swami Vivekananda",
+	},
+	{
+		id: "9",
+		"vibhag id": "99",
+		"book name": "Lorem Ipsum is simply dummy text",
+		author: "Lorem Ipsum",
+	},
+];
+
+export default book_details;
